Create DataWarehouse.R
This is a Warehouse that was designed for merging different internal reports done by several tools, this was a work around meanwhile a more automated solution could be implemented. This code allowed for the creation of a datawarehouse in a shared drive that could be analyzed by several tools like PowerBI or Tableau. code is very easy to implement and it interacts with one of the APIs of the vendors.