Create MappingAgents
This is a code that allows the organization to map in real time what are the schedules and the hierarchy of the agents. This allows to track the movements of the agents, calculate attrition, find errors in hierarchy and other inconsistencies. This is meant to work with a scheduler such as the task scheduler or cron.